Security Advisories: Updates for 11.0.
Various CVEs are now visible at nvd, chgange or add links to point
there, or reve (Bo Yet Public) comments.

Try to ensure that text makes sense with what is in the release
(e.g. remove reference to a sed which is no-longer visible by bumping
the version to use), add 'or later' where we now have a later version,
note that kernels 5.11 and 5.12 are no-longer maintained.