fix multiple definition of sa-11.1.095 in BLFS 11.1 SA
W3 validator says:

An "id" is a unique identifier. Each time this attribute is used in a
document it must have a different value. If you are using this attribute
as a hook for style sheets it may be more appropriate to use classes
(which group elements) than id (which are used to identify exactly one
element).