Tyr to fix w3c validation for latest advisory.
The use of '<...>' in the text pasted from the upstream advisory
breaks things, remove it - people can use the link to the upstream
advisory if they wish to know the 'howto' details.

Validation was very confused about the line number:

Error Line 189, Column 74: required attribute "href" not specified

…eplace HTTPS connections to GitHub with SSH (through git's url.<base>.insteadOf

The attribute given above is required for an element that you've used, but you have omitted it. For instance, in most HTML and XHTML document types the "type" attribute is required on the "script" element and the "alt" attribute is required for the "img" element.