Security Advisories: Issue SA-11.3-008 for Samba.
Note the critical nature of this vulnerability. A proof of concept has
been demonstrated that allows for Bitlocker recovery keys to be
exfiltrated.

However, note that you must have AD DC/LDAP enabled, which is not a
configuration that we describe normally in the book. At least three
users have talked to the lists about it in the past few months though,
so I'm putting this warning out there.