Security Advisories: Issue SA-12.0-068 for systemd.
This vulnerability is not exploitable in a default configuration, but in
case a user has enabled DNSSEC support, let's create an advisory. For
most users, no action will be necessary for this security advisory. For
others, they'll need to rebuild using the sed in either LFS or BLFS.

Since most users who use systemd will need to rebuild it at some point
in BLFS anyway, I'm linking there, but the fix is in LFS as well.