Security Advisories: Issue SA-12.2-097 for VIM.
This will be the last security advisory for 12.2, and I mostly issued it
for my own records, since I'm going to produce a report of security
vulnerability statistics for the 12.3 cycle later.

The next commit will prepare the security advisories for the next cycle,
and then later today new advisories will be filed for Spidermonkey,
Firefox, and Thunderbird.